Dividing drawSphere into two functions: build and render

+++ Lab11/GraphicsProgramming/GraphicsProgramming/FreeCamera.cpp
@@ -12,13 +12,10 @@ FreeCamera::FreeCamera() {
 FreeCamera::~FreeCamera() {}
 
 void FreeCamera::update() {
-	float cosR, cosP, cosY;	//temp values for sin/cos from 
-	float sinR, sinP, sinY;
+	double cosR, cosP, cosY;	//temp values for sin/cos from 
+	double sinR, sinP, sinY;
 	// Roll, Pitch and Yall are variables stored by the FreeCamera
-
 	// handle rotation
-
-
 	// Only want to calculate these values once, when rotation changes, not every frame. 
 	cosY = cosf(Yaw*3.1415 / 180);
 	cosP = cosf(Pitch*3.1415 / 180);
@@ -28,11 +25,8 @@ void FreeCamera::update() {
 	sinR = sinf(Roll*3.1415 / 180);
 
 	//This using the parametric equation of a sphere
-
 	// Calculate the three vectors to put into  glu Lookat
-
 	// Look direction,  position and the up vector 
-
 	// This function could also calculate the right vector
 
 	forward.x = sinY * cosP;
